Change how AddEditVC behaves using the cocoapod "Eureka", now it uses a kind of "reactive" pattern to add rows to a table view that keeps track of all the elements we need to request the user when a transaction is going to be added.

+++ Source/Model/Transaction.swift
@@ -11,14 +11,14 @@ import Foundation
 /// A simple class container for the transactions that are going to
 /// be stored.
 @objc class Transaction: NSObject, Codable {
-    @objc let tID: UInt
+    @objc let tID: Int
     @objc let tName: String
-    @objc let tPrice: Float
+    @objc let tPrice: Double
     @objc let tCurrency: String
     @objc let tDate: String
     @objc let tCats: [String]?
     
-    @objc init(id: UInt, name: String, price: Float,
+    @objc init(id: Int, name: String, price: Double,
                currency: String, date: String, categories: [String]?) {
         self.tID = id
         self.tName = name
